Create Quantifying capsule area per bacterium
This code is written in ImageJ Macro language. It uses FITC image first to calculate the area of the capsule, then to measure the length of the chain. Then the Phase image is used to calculate the area of the bacteria. Capsule area/um of chain is calculated and displayed in the "Results" window at the end of each run.